+++ src/routes/auth/providers/utils.ts
@@ -8,7 +8,8 @@ import {
   PROVIDER_FAILURE_REDIRECT,
   SERVER_URL,
   PROVIDERS,
-  DEFAULT_USER_ROLE
+  DEFAULT_USER_ROLE,
+  DEFAULT_ALLOWED_USER_ROLES,
 } from '@shared/config'
 import { insertAccount, insertAccountProviderToUser, selectAccountProvider } from '@shared/queries'
 import { selectAccountByEmail } from '@shared/helpers'
@@ -96,7 +97,7 @@ const manageProviderStrategy = (
     active: true,
     default_role: DEFAULT_USER_ROLE,
     account_roles: {
-      data: [{ role: DEFAULT_USER_ROLE }]
+      data: DEFAULT_ALLOWED_USER_ROLES.map((role) => ({ role }))
     },
     user: { data: { display_name, avatar_url } },
     account_providers: {