+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - server: `display_name` is attached for email templates
 - server: Added auth middleware to support non-cookie approach for refresh and JWT tokens
 - server: Merge accounts using external OAuth providers.
+- server: Fix default allowed roles for OAuth Providers
 
 ## v2.1.1